backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.8.1
requests-oauthlib==2.0.0
